shift attestation to enclave packages and remove from syft-client
