+++ packages/syft-enclave/pyproject.toml
@@ -10,6 +10,7 @@ dependencies = [
     "syft-client",
     "pydantic-settings>=2.11.0",
     "requests>=2.32.0",
+    "google-auth[pyjwt]>=2.22.0",
 ]
 
 [build-system]
