+++ syft_client/sync/syftbox_manager.py
@@ -828,28 +828,6 @@ class SyftboxManager(BaseModel):
         if sync:
             self.sync()
 
-    def attest_peer(self, peer_email: str):
-        """Verify an enclave peer's attestation by re-reading SYFT_version.json
-        from Drive. Returns None (with an info print) when no token is available;
-        raises AttestationError only when verification of an existing token fails."""
-        from syft_client.sync.version.attestation import verify_attestation_token
-
-        version_info = self.peer_manager.connection_router.read_peer_version_file(
-            peer_email
-        )
-        if version_info is None:
-            print(
-                f"ℹ️  No version file available for peer {peer_email!r}; skipping attestation."
-            )
-            return None
-        if not version_info.attestation_token:
-            print(
-                f"ℹ️  Peer {peer_email!r} has no attestation token "
-                "(not running in a Confidential Space); skipping attestation."
-            )
-            return None
-        return verify_attestation_token(version_info.attestation_token)
-
     def submit_bash_job(
         self,
         user: str,
